V1.12.9

+++ genmonlib/modbus_file.py
@@ -143,7 +143,7 @@ class ModbusFile(ModbusBase):
             if self.StringIsHex(Value):
                 # Not a string, just hex data in a string format
                 for i in range( 0, len(Value), 4):
-                    self.Registers["%04x" % (RegInt + (i / 4))] = Value[i:i+4]
+                    self.Registers["%04x" % (RegInt + int(i / 4))] = Value[i:i+4]
             else:
                 for i in range(0, len(Value), 2):
                     HiByte = ord(Value[i])
@@ -151,7 +151,7 @@ class ModbusFile(ModbusBase):
                         LowByte = 0
                     else:
                         LowByte = ord(Value[i+1])
-                    self.Registers["%04x" % (RegInt + (i / 2))] = "%02x%02x" % (HiByte, LowByte)
+                    self.Registers["%04x" % (RegInt + int(i / 2))] = "%02x%02x" % (HiByte, LowByte)
         return True
 
     #----------  ReadJSONFile  -------------------------------------------------
